[Navigator] Export per-platform NavigatorNavigationBarStyles for consistent styling
This allows for the iOS-style navigation bar on Android and vice versa in order to simplify design. It is entirely optional in that NavigationBars will continue to defauly to their platform-specific style, but you can override it with the `navigationStyles` prop:

```js
<Navigator.NavigationBar
  navigationStyles={Navigator.NavigationBar.StylesIOS}
/>
```

Test Plan: Open the UIExplorer example on iOS and test the plain and breadcrumb Navigator demo.